Overview --- Moving specific definitions to the Utilities following review Also removing the Controller definition as it was replaced with an interface and updating the Handedness enum.
Changes
---
Moves:
* AxisType
* Handedness
* RotationConstraintType

Handedness enum updated following review
