+++ Assets/MixedRealityToolkit/_Core/Definitions/Utilities/Handedness.cs
@@ -3,7 +3,7 @@
 
 using System;
 
-namespace Microsoft.MixedReality.Toolkit.Internal.Definitions
+namespace Microsoft.MixedReality.Toolkit.Internal.Definitions.Utilities
 {
     /// <summary>
     /// The Handedness defines which hand a controller is currently operating in.
@@ -29,6 +29,11 @@ namespace Microsoft.MixedReality.Toolkit.Internal.Definitions
         /// The controller is identified as being either left and/or right handed.
         /// </summary>
         Both = Left | Right,
+        /// <summary>
+        /// Global catchall, used to map actions to any controller (provided the controller supports it)
+        /// *Note, by default the specific hand actions will override settings mapped as both
+        /// </summary>
+        Any = None | Left | Right | Other,
         /// <summary>
         /// Reserved, for systems that provide alternate hand state.
         /// </summary>
